requirements.txt
pydantic>=2.0
python-dotenv
requests>=2.28
PyYAML>=6.0
pydantic-settings>=2.0
types-PyYAML>=6.0

# 개발·테스트 (Phase 1-1)
pytest>=7.0
pytest-cov>=4.0
responses>=0.25.0
types-requests>=2.32.0
ruff>=0.6.0
pre-commit>=3.0
typer>=0.12.0